+++ Public/New-sqmOlaMaintenanceJobs.ps1
@@ -319,7 +319,7 @@ function New-sqmOlaMaintenanceJobs
 					$op = Get-DbaAgentOperator @connParams -Operator $OperatorName -ErrorAction SilentlyContinue
 					if ($op)
 					{
-						Set-DbaAgentJob @connParams -Job $Name -OperatorToEmail $OperatorName -EmailLevel OnFailure -ErrorAction SilentlyContinue | Out-Null
+						Set-DbaAgentJob @connParams -Job $Name -EmailOperator $OperatorName -EmailLevel OnFailure -ErrorAction SilentlyContinue | Out-Null
 					}
 					else
 					{