+++ Public/New-sqmOlaSysDbBackupJob.ps1
@@ -288,7 +288,7 @@ EXECUTE master.dbo.DatabaseBackup
 				$op = Get-DbaAgentOperator @connParams -Operator $OperatorName -ErrorAction SilentlyContinue
 				if ($op)
 				{
-					Set-DbaAgentJob @connParams -Job $effJobName -OperatorToEmail $OperatorName -EmailLevel OnFailure -ErrorAction SilentlyContinue | Out-Null
+					Set-DbaAgentJob @connParams -Job $effJobName -EmailOperator $OperatorName -EmailLevel OnFailure -ErrorAction SilentlyContinue | Out-Null
 				}
 				else
 				{